Devs2

+++ app/Views/auth/profile.view.php
@@ -1,6 +1,6 @@
 <?php require views_path('partials/header');?>
 
-	<div class="container-fluid border col-lg-5 col-md-6 mt-5 p-4" >
+	<div class="container-fluid border col-lg-6 col-md-6 mt-5 p-4" >
 		
 		<?php if(is_array($row)):?>
 
@@ -10,15 +10,17 @@
 			</center>
 			<br>
 
-			<table class="table table-hover table-striped">
+			<table class="table table-hover table-borderless" style="background:#F7F7F7;">
 				<tr>
-					<td rowspan="6">
-						<center> <img src="<?=crop($row['image'],400,$row['gender'])?>" style="width: 100%;max-width:200px;" > </center>
+					<td rowspan="7">
+						<center> <img src="<?=crop($row['image'],400,$row['gender'])?>" style="width: 100%;max-width:250px;" > </center>
 					</td>
 				</tr>
-
 				<tr>
-					<th>Username</th><td><?=$row['username']?></td>
+					<th>Employee ID</th><td><?=$row['userid']?></td>
+				</tr>
+				<tr>
+					<th>Name</th><td><?=$row['username']?></td>
 				</tr>
 				<tr>
 					<th>Email</th><td><?=$row['email']?></td>
@@ -35,18 +37,31 @@
 				
 			</table>
 			<br>
-			<a href="index.php?pg=edit-user&id=<?=$row['id']?>">
-				<button type="button" class="btn btn-primary">Edit</button>
+			<?php if(auth('role') == "Cashier"): ?>
+			<a href="index.php?pg=home">
+				<button type="button" class="btn btn-secondary">Back</button>
 			</a>
+			<?php else: ?>
+			<a href="index.php?pg=admin&tab=users">
+				<button type="button" class="btn btn-secondary">Back</button>
+			</a>
+			<?php endif; ?>
+			<?php if(Auth::access('Admin') || ($row && $row['id'] == Auth::get('id'))): ?>
+				<a href="index.php?pg=edit-user&id=<?=$row['id']?>">
+					<button type="button" class="btn btn-primary">Edit</button>
+				</a>
 
-			<a href="index.php?pg=delete-user&id=<?=$row['id']?>" class="float-end">
-				<button type="button" class="btn btn-danger">Delete</button>
-			</a>
+				<a href="index.php?pg=delete-user&id=<?=$row['id']?>" class="float-end">
+					<button type="button" class="btn btn-danger">Delete</button>
+				</a>
+			<?php endif;?>
 
 
 		<?php else:?>
 			<div class="alert alert-danger text-center">That user was not found!</div>
-
+			<a href="index.php?pg=admin&tab=users">
+				<button type="button" class="btn btn-secondary">Back</button>
+			</a>
 		<?php endif;?>
 	</div>
